requirements.txt
discord.py==2.4.0
python-dotenv==1.0.1
flask==3.0.3
colorama==0.4.6
requests==2.31.0
